Re-introduce education stage step for ETAs
We removed the education stage step for ETAs but this led to a confusing
subject you'd like to teach step where they had to select 'primary' as their
subject if they wanted to teach at primary stage.

Re-introduce the education stage step for all candidates. Remove 'primary'
from the subject you'd like to teach step as an option (as we only display
this step to those interested in teaching secondary). Update the helper text
for the subject you'd like to teach step.