Add Google Optimize for split testing
We are split testing the TTA sign up on the GiT website with the Get an adviser
service; in order for the test to work correctly the optimize script needs to
be on both variants. It's already on the GiT website, but this will pull it
into the Get an adviser service home page.